Add Angular (#1)
Parent story: https://github.com/fenya123/bingin/issues/18

In the scope of this task we need to configure angular with minimal setup required at this moment. Also as a part of this task we're going to add some configuration for `npm` and `nvm`. And update README.md to expose our project setup guide.