Add and configure ESLint (#5)
Parent story: https://github.com/fenya123/bingin/issues/9

As a part of adding linters configuration story we need to add and configure eslint.

We try to work with strictest configuration as possible, so we have to enable all available linter rules, and disable only those we don't actually need.

In the scope of this task we need to add and configure eslint rules and refactor existing code to comply with the rules.