Add and configure ESLint plugins (#7)
Parent story: https://github.com/fenya123/bingin/issues/9

As a part of adding linters configuration story we need to add and configure eslint plugins after we've already added and configured eslint itself.

We try to work with strictest configuration as possible, so we have to enable all available linter rules, and disable only those we don't actually need. But now we don't have enough time to attentively go through all available plugins and their rules, so we decided to add only those which looks useful at a first glance, enable all rules, and configure them later at the time of development.

We're going to add following eslint plugins:
- [@angular-eslint/eslint-plugin](https://github.com/angular-eslint/angular-eslint/tree/main/packages/eslint-plugin)
- [@angular-eslint/eslint-plugin-template](https://github.com/angular-eslint/angular-eslint/tree/main/packages/eslint-plugin-template)
- [eslint-plugin-no-comments](https://github.com/wisniewski94/eslint-plugin-no-comments)
- [eslint-plugin-perfectionist](https://github.com/azat-io/eslint-plugin-perfectionist)

In the scope of this task we need to add and configure different eslint plugins and refactor existing code to comply with the rules.