Add typography (#12)
Parent story: https://github.com/fenya123/bingin/issues/25

As a part of setting up basic frontend stuff we should develop and implement typography system which we will use in our app.

In the scope of this ticket we have to setup font family and font variations we're going to use add css classes which will contain according typography rules.